go: split predicates to files
Move each predicate into it's own file to match the test setup. Find
this a bit easier to keep things organized. Still need to rename the
functions to match the target API.
